✨ feat: full-bleed grouped style on the three browse lists
Switch Home (scenario list + paused resume card), Shared Scenarios, and Past
Results to the new `.grouped` PasturaSection/PasturaCard form (#731): white
bands reach both screen edges, no radius, top+bottom hairlines only, with
text-inset row separators.

Section headers carry their own screen-edge inset now that `.grouped` zeroes
the section's horizontal margin. SharedScenarios' offline banner and category
chips keep their independent insets (not full-bleed). The category-chip
helpers move to SharedScenariosListView+CategoryChips.swift to stay within the
type_body_length budget.

Part of #731
