+++ Pastura/Pastura/Views/Home/HomePausedCard.swift
@@ -22,7 +22,9 @@ struct HomePausedCard: View {
   }
 
   var body: some View {
-    PasturaCard {
+    // .grouped so the resume card reads as the same full-bleed band as the
+    // scenario list stacked beneath it on Home (#731).
+    PasturaCard(style: .grouped) {
       VStack(alignment: .leading, spacing: 11) {
         Text(summary.name)
           .font(.headline)
